Moved everything out of their packages to set up for branch-A-Packages
