+++ src/main/java/duke/Event.java
@@ -1,4 +1,4 @@
-package duke.task;
+package duke;
 
 public class Event extends Task {
