+++ src/main/java/duke/StringTools.java
@@ -1,4 +1,4 @@
-package duke.command;
+package duke;
 
 import java.util.Arrays;
